phidata/requirements.txt
python-dotenv
phidata==2.7.5
groq
yfinance
duckduckgo-search